Fix remaining MSVC size_t to int conversion warnings
- Add static_cast<int>() for size_t loop indices in map_generator.cpp
- Add static_cast<int>() for vector sizes in map_validator.cpp
- Add static_cast<int>() for items.size() in inventory.cpp
- Add static_cast<char>() for slot letter calculation in inventory_renderer.cpp
- Add static_cast<float>() for sqrt results in test_spawn_manager.cpp
- Add static_cast<int>() for getMonsters().size() in test_spawn_manager.cpp

These conversions are safe as the values are guaranteed to be within int range.
This should fix the Windows CI build failures with MSVC treating warnings as errors.

+++ src/inventory.cpp
@@ -164,7 +164,7 @@ std::vector<Item*> Inventory::getAllItems() const {
 
 int Inventory::findEmptySlot() const {
     if (items.size() < static_cast<size_t>(capacity)) {
-        return items.size();
+        return static_cast<int>(items.size());
     }
     return -1;
 }
